backend: rearrange files tructure
