+++ backend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "master-word-api",
-  "version": "0.2.1",
+  "version": "0.3.1",
   "description": "NodeJS backend for the Master Word app",
   "main": "index.js",
   "scripts": {
