+++ backend/src/constants.ts
@@ -1,0 +1,2 @@
+export const MAX_ATTEMTPS = 8 as const;
+export const WORD_LENGTH = 5 as const;
