+++ backend/src/router/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+import { default as gameRoutes } from './game.routes';
+export { gameRoutes };
